Move charts to their own component
