+++ src/scss/components/PerfCharts.scss
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
+//
+// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
+.charts-container {
+    display: grid;
+    grid-template-columns: 1fr minmax(min-content, 750px) 1fr;
+    position: relative;
+    gap: 60px;
+}
