--- src/scss/components/Performance.scss
@@ -1,16 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-//
-// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
-
-.performance {
-    .button-container {
-        margin-bottom: 20px;
-    }
-
-    .charts-container {
-        display: grid;
-        grid-template-columns: 1fr minmax(min-content, 750px) 1fr;
-        position: relative;
-        gap: 60px;
-    }
-}
